bkmk/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.0.7
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.1.1
django-extensions==2.2.9
djangorestframework==3.11.0
importlib-metadata==1.6.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
